requirements.txt
Django==5.2.12
djangorestframework==3.17.1
psycopg2-binary==2.9.11
requests==2.33.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==2.2.6
xgboost==3.2.0
